refactor for better maintainability
